save cached value from memcached to local cache it is exists
Before: when a value from memcached was taken it was not set to the
local cache. Therefore, the script always called memcached to get the
value but never the local cache.

After: if a value exists in memcached it is saved to a local cache to be
used later (not from memcached because it is slower).